Switch history from btree to queue.
Previously, history was stored as a DB_HASH database with a DB_BTREE secondary
index on time added, to allow history expiry.  Unfortunately DB_TREE is very
slow, and NTS would spend ~60% of its time updating the database.

The database is now a DB_QUEUE appended in order of article receipt, with a
DB_HASH secondary index on message-id to allow lookups.  Expiry simply starts
at the first record in the database and continues until it finds an article
too new to expire.

In benchmarking, this increased article receipt throughput from 2,000/sec to
8,000/sec, a 5x improvement.