Client article buffering.
Add a configuration option 'peer::article-buffer' which, if set, causes
us to respond with 239 (accepted) if we're waiting for an article to
process after TAKETHIS and the client sends another command.  This has
the potential to increase throughput; the downside is that if we crash,
any buffered articles will be lost.  (So, try not to crash.)